Local max counter as translation unit
